backend/requirements.txt
Pillow==10.1.0
Django==3.2
djangorestframework==3.12.4
python-dotenv==1.0.0
djoser==2.2.2
django-filter==23.5
subscribe==0.6.1